automatically discover tests and run them in parallel
