+++ .travis.yml
@@ -19,7 +19,7 @@ before_install:
   - conda install mamba
   - conda info -a
 install:
-  - mamba create -q -n test-environment -c flatsurf sagelib=$SAGE_VERSION pip pytest surface-dynamics ipywidgets scipy matplotlib-base gap-defaults $CONDA_PACKAGES
+  - mamba create -q -n test-environment -c flatsurf sagelib=$SAGE_VERSION pip pytest pytest-xdist surface-dynamics ipywidgets scipy matplotlib-base gap-defaults $CONDA_PACKAGES
   - conda activate test-environment
   - pip install $PIP_PACKAGES
   - pip install --verbose --no-index .
@@ -27,4 +27,4 @@ install:
   - python -c 'import cppyy' || true
 script:
   - sage -tp --force-lib --long --optional=$SAGE_OPTIONALS flatsurf doc
-  - ./test/run_tests.py
+  - pytest -n auto
