+++ test/gl2r_orbit_closure/test_veech_n_gons_and_double_n_gons.py
@@ -1,7 +1,14 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 r"""
-Calta-McMullen Veech surfaces in H(2)
+Veech 2n-gons and double n-gons
+
+These two families of examples are from the article of Veech "Teichmüller
+curves in moduli space, Eisenstein series and an application to triangular
+billiards" (1989). These are examples of what now is called a Veech surface.
+These surface satisfies the so-called Veech dichotomy: in any direction of a
+saddle connection, the flow is completely periodic and the cylinders have
+commensurable moduli ("parabolic direction").
 """
 ######################################################################
 # This file is part of sage-flatsurf.
@@ -25,19 +32,25 @@ Calta-McMullen Veech surfaces in H(2)
 import sys
 import pytest
 
-from sage.all import polygen, NumberField, AA, QQ
+import sage.all
 from flatsurf import translation_surfaces, GL2ROrbitClosure
 
-@pytest.mark.parametrize("w,h,t,e", [(2,1,0,0), (3,1,0,0), (3,1,0,1), (4,1,0,1),
-               (4,1,0,2), (3,2,0,0), (4,2,1,0), (4,2,0,1), (4,2,1,1)])
-def test_H2(w,h,t,e):
-    S = translation_surfaces.mcmullen_genus2_prototype(w,h,t,e)
+@pytest.mark.parametrize("n,bound", [(4,4),(5,4),(6,4),(7,4),(8,4)])
+def test_veech_2n_gon(n, bound):
+    S = translation_surfaces.veech_2n_gon(n)
     O = GL2ROrbitClosure(S)
-    for d in O.decompositions(5, 50):
+    for d in O.decompositions(bound):
         assert d.parabolic()
 #        assert d.decomposition.cylinder_diagram()[0].stratum() == O.surface.stratum()
         O.update_tangent_space_from_flow_decomposition(d)
     assert O.dimension() == O.absolute_dimension() == 2
 
-if __name__ == '__main__': sys.exit(pytest.main(sys.argv))
-
+@pytest.mark.parametrize("n,bound", [(3,4),(5,4),(7,4),(9,4)])
+def test_veech_double_n_gon(n, bound):
+    S = translation_surfaces.veech_double_n_gon(n)
+    O = GL2ROrbitClosure(S)
+    for d in O.decompositions(bound):
+        assert d.parabolic()
+#        assert d.decomposition.cylinder_diagram()[0].stratum() == O.surface.stratum()
+        O.update_tangent_space_from_flow_decomposition(d)
+    assert O.dimension() == O.absolute_dimension() == 2
